Small upstream-parity fixes and dead-code cleanup
- internal/ppu/merge.go: SpritePixel.Color() now masks 0xFF instead of
  0x1FF. Writer at sprite.go:11 stores 8-bit color (`color & 0xFF`),
  matching upstream ppu.hh:345 (`u8 color : 8`). Bit 8 was always zero
  in practice but the 9-bit accessor was a latent footgun.
- internal/scheduler/scheduler.go: TimestampTarget lazily pops
  cancelled events sitting at the heap top. Upstream Cancel() fixes
  the heap in-place; Go's container/heap can't do that without extra
  index bookkeeping, so popping here keeps RemainingCycleCount honest
  for the halt loop.
- internal/timer/timer.go: change ticksMask and Channel.mask from int64
  to int to match upstream timer.hh (`int mask`). Cast at the use site
  in startChannel for the int64 scheduler.Now() &.
- internal/bus/bus.go: drop the dead Bus.step() wrapper (never called
  — all sites use stepAt) and the `var _ = false` gofmt-pinning
  comment-collapse workaround.
- internal/arm/arithmetic.go: drop the unused bus import and the
  `var _ = bus.AccessNonsequential` keep-alive marker; the import
  was a leftover from a prior refactor.

+++ internal/timer/timer.go
@@ -23,7 +23,7 @@ const (
 
 var (
 	ticksShift = [4]int{0, 6, 8, 10}
-	ticksMask  = [4]int64{0, 0x3F, 0xFF, 0x3FF}
+	ticksMask  = [4]int{0, 0x3F, 0xFF, 0x3FF}
 )
 
 type pending struct {
@@ -47,7 +47,7 @@ type channel struct {
 
 	running          bool
 	shift            int
-	mask             int64
+	mask             int
 	timestampStarted int64
 
 	eventOverflow scheduler.EventID
@@ -201,7 +201,7 @@ func (t *Timer) onControlWritten(chanID uint64) {
 	ch.mask = ticksMask[ch.control.frequency]
 
 	if ch.control.enable {
-		prescalerOffset := t.scheduler.Now() & ch.mask
+		prescalerOffset := t.scheduler.Now() & int64(ch.mask)
 		if enablePrevious {
 			if !ch.control.cascade {
 				t.startChannel(ch, prescalerOffset)
